Added Request Logging. (Very Verbose)
Added a logging flag that can be passed to the server command through
either -L or --logging

Doing so results in the request at the middleware level to be logged to
the console. It's fairly basic, but shows a lot of information. It would
be good to direct this output into a file or something else.